+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- New `Unique` and `Unique2` scalar iterators to iterate over unique values.
+
 ## 1.0.0 - 2025-05-24
 
 This is the first stable release of *gloop*! :tada: